mag_6895 to 7009 Merged NEW REGULATION

+++ rx_digiprint_gui/Models/CleafRoll.cs
@@ -64,7 +64,7 @@ namespace RX_DigiPrint.Models
         public void send_quality(bool? quality)
         {
             TcpIp.SValue value = new TcpIp.SValue();
-            value.no = (UInt32)No;
+            value.no = No;
             switch(quality)
             {
                 case false: value.value=1; break;
